fix(fs_gateway): close remote-mount review findings
Blob access is scoped to the token's bucket: AllocateBlobGuid writes an
@fs-owner marker into the bucket's NSS root, data requests carry the inode
key, and the gateway proves ownership from a per-bucket cache, the marker,
or the layout at that key (the only proof for S3-created blobs). Read
paths never write NSS. Sweep hints are re-derived in the sweep worker
after the grace: whole-blob teardown needs the key unpublished or a
committed @ovr-gc marker, and victims that resolve live are dropped.
DeleteBlock and DeleteBlobBlocks leave the protocol.

The wire header stores the command as i32 and servers validate it with
Command::try_from (an invalid discriminant was UB); with every field a
primitive the header is a plain struct with derived Pod, no hand-written
unsafe impls. Bodies are capped at 1 MiB and blocks at one shard;
in-flight requests are bounded per connection and per process. Session
tokens are per connection so mounts behind a load balancer do not
thrash. Rename and unlink are not transport-retried on the client hop.
The disk cache mirrors a written block only when the write-once put
actually stored it. Deploy builds package fs_gateway and
fractalbits-mount via the isolated compio build.

Verified: just run-tests fs-server (121 passed), just fstest (4333
passed), just precheckin, untar-bench and overwrite-bench at parity
with main.

+++ crates/common/rpc_clients/nss/src/rpc.rs
@@ -38,7 +38,7 @@ impl RpcClient {
         let mut header = MessageHeader::default();
         let request_id = self.gen_request_id();
         header.id = request_id;
-        header.command = Command::PutInode;
+        header.command = Command::PutInode as i32;
         header.size = (size_of::<MessageHeader>() + body.encoded_len()) as u32;
         header.retry_count = retry_count as u8;
         header.set_trace_id(trace_id);
@@ -89,7 +89,7 @@ impl RpcClient {
         let mut header = MessageHeader::default();
         let request_id = self.gen_request_id();
         header.id = request_id;
-        header.command = Command::PutInodeCas;
+        header.command = Command::PutInodeCas as i32;
         header.size = (size_of::<MessageHeader>() + body.encoded_len()) as u32;
         header.retry_count = retry_count as u8;
         header.set_trace_id(trace_id);
@@ -130,7 +130,7 @@ impl RpcClient {
         let mut header = MessageHeader::default();
         let request_id = self.gen_request_id();
         header.id = request_id;
-        header.command = Command::GetInode;
+        header.command = Command::GetInode as i32;
         header.size = (size_of::<MessageHeader>() + body.encoded_len()) as u32;
         header.retry_count = retry_count as u8;
         header.set_trace_id(trace_id);
@@ -182,7 +182,7 @@ impl RpcClient {
         let mut header = MessageHeader::default();
         let request_id = self.gen_request_id();
         header.id = request_id;
-        header.command = Command::ListInodes;
+        header.command = Command::ListInodes as i32;
         header.size = (size_of::<MessageHeader>() + body.encoded_len()) as u32;
         header.retry_count = retry_count as u8;
         header.set_trace_id(trace_id);
@@ -246,7 +246,7 @@ impl RpcClient {
         let mut header = MessageHeader::default();
         let request_id = self.gen_request_id();
         header.id = request_id;
-        header.command = Command::DeleteInode;
+        header.command = Command::DeleteInode as i32;
         header.size = (size_of::<MessageHeader>() + body.encoded_len()) as u32;
         header.retry_count = retry_count as u8;
         header.set_trace_id(trace_id);
@@ -283,7 +283,7 @@ impl RpcClient {
         let mut header = MessageHeader::default();
         let request_id = self.gen_request_id();
         header.id = request_id;
-        header.command = Command::CreateRootInode;
+        header.command = Command::CreateRootInode as i32;
         header.size = (size_of::<MessageHeader>() + body.encoded_len()) as u32;
         header.retry_count = retry_count as u8;
         header.set_trace_id(trace_id);
@@ -344,7 +344,7 @@ impl RpcClient {
         let mut header = MessageHeader::default();
         let request_id = self.gen_request_id();
         header.id = request_id;
-        header.command = Command::DeleteRootInode;
+        header.command = Command::DeleteRootInode as i32;
         header.size = (size_of::<MessageHeader>() + body.encoded_len()) as u32;
         header.retry_count = retry_count as u8;
         header.set_trace_id(trace_id);
@@ -386,7 +386,7 @@ impl RpcClient {
         let mut header = MessageHeader::default();
         let request_id = self.gen_request_id();
         header.id = request_id;
-        header.command = Command::Rename;
+        header.command = Command::Rename as i32;
         header.size = (size_of::<MessageHeader>() + body.encoded_len()) as u32;
         header.retry_count = retry_count as u8;
         header.set_trace_id(trace_id);
@@ -445,7 +445,7 @@ impl RpcClient {
         let mut header = MessageHeader::default();
         let request_id = self.gen_request_id();
         header.id = request_id;
-        header.command = Command::Rename;
+        header.command = Command::Rename as i32;
         header.size = (size_of::<MessageHeader>() + body.encoded_len()) as u32;
         header.retry_count = retry_count as u8;
         header.set_trace_id(trace_id);
